Refactor code generator structure
